feat: refactor functions
