+++ backend/product-service/get-products-by-id.ts
@@ -1,28 +1,17 @@
-import { products as mockProducts } from "./mocks/products";
+import { httpResponse } from "./common/http-responses";
+import { products } from "./mocks/products";
 
-const headers = {
-  "Content-Type": "application/json",
-  "Access-Control-Allow-Headers": "Content-Type",
-  "Access-Control-Allow-Origin": "https://d2qorf0xmzna5y.cloudfront.net",
-  "Access-Control-Allow-Methods": "OPTIONS,POST,GET",
-};
-
-export const handler = async (event: any) => {
-  const products = mockProducts;
+export const handler = async (event: {
+  pathParameters: { productId: string };
+}) => {
   const productId = event.pathParameters.productId;
   const product = products.find((p) => p.id === productId);
 
   if (!product) {
-    return {
-      statusCode: 404,
-      headers,
-      body: JSON.stringify({ message: "Product not found" }),
-    };
+    return httpResponse(404, {
+      message: "Product not found",
+    });
   }
 
-  return {
-    statusCode: 200,
-    headers,
-    body: JSON.stringify(product),
-  };
+  return httpResponse(200, product);
 };
