+++ backend/product-service/get-products-list.ts
@@ -1,15 +1,6 @@
+import { httpResponse } from "./common/http-responses";
 import { products as mockProducts } from "./mocks/products";
 
-export const handler = async (event: any) => {
-  const products = mockProducts;
-  return {
-    statusCode: 200,
-    headers: {
-      "Content-Type": "application/json",
-      "Access-Control-Allow-Headers": "Content-Type",
-      "Access-Control-Allow-Origin": "https://d2qorf0xmzna5y.cloudfront.net",
-      "Access-Control-Allow-Methods": "OPTIONS,POST,GET",
-    },
-    body: JSON.stringify(products),
-  };
+export const handler = async () => {
+  return httpResponse(200, mockProducts);
 };
